+++ src/Enums/Alternative.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace HiFolks\Statistics\Enums;
+
+enum Alternative: string
+{
+    case TwoSided = 'two-sided';
+    case Greater = 'greater';
+    case Less = 'less';
+}